Add focus on first taskItem upon loading
