+++ src/main/java/vimification/taskui/TaskListViewCell.java
@@ -1,0 +1,20 @@
+package vimification.taskui;
+
+import javafx.scene.control.ListCell;
+import vimification.model.task.Task;
+
+/**
+ * Custom {@code ListCell} that displays the graphics of a {@code Task} using a {@code TaskCard}.
+ */
+class TaskListViewCell extends ListCell<Task> {
+    @Override
+    protected void updateItem(Task task, boolean empty) {
+        super.updateItem(task, empty);
+        if (empty || task == null) {
+            setGraphic(null);
+            setText(null);
+            return;
+        }
+        setGraphic(new TaskCard(task, getIndex() + 1).getRoot());
+    }
+}
